agents/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
bedrock-agentcore==0.1.2
bedrock-agentcore-starter-toolkit
boto3
botocore
pytest
